Use render_parent to render new issue template
Before this commit the plugin used copy-paste method to provide modified
templates. The new issue template was just a copy of a original redmine
template with a few strings modified.

This led to some problems. First of all, the plugin failed in some cases
with some newer versions of redmine. For instance, with redmine 3.3.8
it happened to respond 500 result when the "New issue" page was access
outside of a project scope. This can be tested by clicking at the
"New issue" link at http://redmine-host-name/issues path.

In order to solve this problem I used a different approach of modifying
the original view templates. This approach is described at the redmine
boards [1]. The approach implies usage of the 'render_parent' gem which
allows you to render an original view when you override it by a plugin.
The original view render is then modified with a help of Nokogiri. Thus
I created a new new issue view template which loads the original
template and replaces the original watcher form with a new one. It also
remove a JS script snippet from the original page which seem unnecessary
and break things when is used with Nokogiri.

[1] https://www.redmine.org/boards/3/topics/33949